requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pytest>=8.0.0
tavily-python>=0.5.0
requests>=2.31.0
trafilatura>=2.0.0
